refactor: #47 강아지 생성 로직 수정

+++ src/main/java/team9/ddang/dog/service/request/UpdateDogServiceRequest.java
@@ -3,6 +3,7 @@ package team9.ddang.dog.service.request;
 import team9.ddang.global.entity.Gender;
 import team9.ddang.dog.entity.IsNeutered;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record UpdateDogServiceRequest(
@@ -10,7 +11,7 @@ public record UpdateDogServiceRequest(
         String name,
         String breed,
         LocalDate birthDate,
-        Integer weight,
+        BigDecimal weight,
         Gender gender,
         String profileImg,
         IsNeutered isNeutered,
